+++ src/lib/mina/v2/index.ts
@@ -8,3 +8,5 @@ export * from './state.js';
 export * from './transaction.js';
 export * from './views.js';
 export * from './zkapp-logic.js';
+export * from './dsl/mina-program.js';
+export * from './currency.js';